Support compact block in data_man.
